Added another describe
Before, when running the KoanRunner, the 'in keyword' describe would appear
before the two expects above it if it hadn't been filled out yet, which isn't
the expected functionality. By adding another describe block, this issue goes
away and we also have more description of those two tests.

It might be a good idea to add a fourth describe block to the final two tests
on this page, just to be consistent. Those would be adding and deltion, as well
as prototypes. Just an idea.