Switched msg to easier to evaluate values array
I found the falsetruefalse msg to be confusing - it is really strang to
have that as a joined string, even though the result makes sense. By having an
array instead, this should be a bit easier for the beginner to parse, while still
communicating the point of the expect statement.